+++ packages/material_ui/test/bottom_sheet_test.dart
@@ -2463,9 +2463,8 @@ void main() {
                   child: ElevatedButton(
                     child: const Text('Press me'),
                     onPressed: () {
-                      Scaffold.of(
-                        context,
-                      ).showBottomSheet((BuildContext context) => const Text('BottomSheet'));
+                      Scaffold.of(context)
+                          .showBottomSheet((BuildContext context) => const Text('BottomSheet'));
                     },
                   ),
                 );
@@ -2589,9 +2588,8 @@ void main() {
                   child: ElevatedButton(
                     child: const Text('Press me'),
                     onPressed: () {
-                      Scaffold.of(
-                        context,
-                      ).showBottomSheet((BuildContext context) => const Text('BottomSheet'));
+                      Scaffold.of(context)
+                          .showBottomSheet((BuildContext context) => const Text('BottomSheet'));
                     },
                   ),
                 );