+++ packages/material_ui/test/will_pop_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:material_ui/material_ui.dart';
+
 import 'finders.dart';
 
 bool willPopValue = false;